Separate declaration and initialization into distinct concepts in Claro.
Now, support type checking of all kinds of situations including the finicky empty list assignment: `l = [];`, we now have var declarations separate from assignments.

Also added the additional `type` Stmt to Claro which for now is a hacked, lazy version of what it will end up being long term. For now, it simply prints a string repr of the type of the given expr which is known at compile time. This is not as strong as what we'll want long term which is to have the ability to have the meta type Type included in the runtime of the language itself as a builtin type in order to allow for high order programming. This is skipped for now since I'd want to simply have the JavaSource impl depend on the same Types.java file that the compiler impl depends on, but I genuinely cannot figure out how to have our genrule use javac on multiple java files yet....

+++ src/java/com/claro/examples/calculator_example/intermediate_representation/AddNumericExpr.java
@@ -10,7 +10,7 @@ import com.google.common.collect.ImmutableSet;
 
 public class AddNumericExpr extends NumericExpr {
 
-  private static final ImmutableSet<Type> SUPPORTED_ADD_OPERAND_TYPES = ImmutableSet.of(Types.INTEGER, Types.DOUBLE);
+  private static final ImmutableSet<Type> SUPPORTED_ADD_OPERAND_TYPES = ImmutableSet.of(Types.INTEGER, Types.FLOAT);
   private final String TYPE_SUPPORT_ERROR_MESSAGE =
       "Internal Compiler Error: Currently `+` is not supported for types other than Integer and Double.";
 
@@ -27,8 +27,8 @@ public class AddNumericExpr extends NumericExpr {
     assertSupportedExprType(lhs, SUPPORTED_ADD_OPERAND_TYPES);
     assertSupportedExprType(rhs, SUPPORTED_ADD_OPERAND_TYPES);
 
-    if (lhs.equals(Types.DOUBLE) || rhs.equals(Types.DOUBLE)) {
-      return Types.DOUBLE;
+    if (lhs.equals(Types.FLOAT) || rhs.equals(Types.FLOAT)) {
+      return Types.FLOAT;
     } else {
       return Types.INTEGER;
     }
