+++ src/java/com/claro/examples/calculator_example/intermediate_representation/FloatTerm.java
@@ -18,7 +18,7 @@ final public class FloatTerm extends Term {
 
   @Override
   protected Type getValidatedExprType(ScopedHeap unusedScopedHeap) {
-    return Types.DOUBLE;
+    return Types.FLOAT;
   }
 
   @Override
