+++ src/java/com/claro/examples/calculator_example/intermediate_representation/LessThanBoolExpr.java
@@ -19,7 +19,7 @@ public class LessThanBoolExpr extends BoolExpr {
 
   @Override
   protected ImmutableSet<Type> getSupportedOperandTypes() {
-    return ImmutableSet.of(Types.INTEGER, Types.DOUBLE);
+    return ImmutableSet.of(Types.INTEGER, Types.FLOAT);
   }
 
   @Override
