+++ src/java/com/claro/examples/calculator_example/intermediate_representation/LessThanOrEqualToBoolExpr.java
@@ -15,7 +15,7 @@ public class LessThanOrEqualToBoolExpr extends BoolExpr {
 
   @Override
   protected ImmutableSet<Type> getSupportedOperandTypes() {
-    return ImmutableSet.of(Types.INTEGER, Types.DOUBLE);
+    return ImmutableSet.of(Types.INTEGER, Types.FLOAT);
   }
 
   @Override
