+++ src/java/com/claro/examples/calculator_example/intermediate_representation/SubtractNumericExpr.java
@@ -11,7 +11,7 @@ import com.google.common.collect.ImmutableSet;
 public class SubtractNumericExpr extends NumericExpr {
 
   private static final ImmutableSet<Type> SUPPORTED_SUBTRACT_OPERAND_TYPES =
-      ImmutableSet.of(Types.INTEGER, Types.DOUBLE);
+      ImmutableSet.of(Types.INTEGER, Types.FLOAT);
   private static final String TYPE_SUPPORT_ERROR_MESSAGE =
       "Internal Compiler Error: Currently `-` is not supported for types other than Integer and Double.";
 
@@ -28,8 +28,8 @@ public class SubtractNumericExpr extends NumericExpr {
     assertSupportedExprType(lhs, SUPPORTED_SUBTRACT_OPERAND_TYPES);
     assertSupportedExprType(rhs, SUPPORTED_SUBTRACT_OPERAND_TYPES);
 
-    if (lhs.equals(Types.DOUBLE) || rhs.equals(Types.DOUBLE)) {
-      return Types.DOUBLE;
+    if (lhs.equals(Types.FLOAT) || rhs.equals(Types.FLOAT)) {
+      return Types.FLOAT;
     } else {
       return Types.INTEGER;
     }
